REVERT: angler: fix first audio track
*It causes weird issues when the user holds their phone wrong*

It seems that the first time an audio track was played was when the
audio clibration data is sent down to the hw. However it wasn't being
properly applied or we missed it and that first track will always sound
awful.

The speaker protection extension fixes this by pre-loading the
calibration data before the first track is played.

Ticket: OSS CYAN-7486
Change-Id: Ibb1ace9dfbf2c1dec4844d1eedd8843889415792
Signed-off-by: Roman Birg <roman@cyngn.com> (reverted from commit 6b34f078b4cb1de9ac704bb8d0cf3e135b7d180d)